hitechdemo/requirement.txt
Django==2.1.1
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.8.2
mongoengine==0.15.3
pymongo==3.7.1
pytz==2018.5
six==1.11.0
